教育用課題のサンプル修正
* informationテーブルの型変更
* 更新時に掲載日を変更しないように修正
* モーダルを閉じるタイミングを変更
* backendでエラー発生時にfrontで表示するメッセージを変更

+++ backend/db/information.js
@@ -1,4 +1,4 @@
-const { STRING, DATE, TINYINT } = require("sequelize");
+const { STRING, DATE, SMALLINT } = require("sequelize");
 
 module.exports.getInformationModel = function (db) {
   return db.define(
@@ -6,7 +6,7 @@ module.exports.getInformationModel = function (db) {
     {
       no: {
         field: "no",
-        type: TINYINT,
+        type: SMALLINT,
         primaryKey: true,
         allowNull: false,
       },
